updated logging and build mode, added watchmode type for shell
